# go/api/handlers/activity.go
package handlers

import (
	"challenge-flutter-go/api/errorHandlers"
	"challenge-flutter-go/api/requests"
	"challenge-flutter-go/api/responses"
	"challenge-flutter-go/api/utils"
	"challenge-flutter-go/logger"
	"challenge-flutter-go/models"
	"challenge-flutter-go/repository"
	"fmt"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type ActivityHandler struct {
	Repository     repository.ActivityRepository
	TripRepository repository.TripRepository
}

// @Summary		Get all activities from a trip
// @Description	Get all activities associated with the trip
// @Tags			activity
// @Accept		json
// @Produce		json
// @Security		BearerAuth
// @Param		id	path		string	true	"ID of the trip"
// @Success		200		{object}	[]responses.ActivityResponse
// @Failure		400		{object}	error
// @Failure		401		{object}	error
// @Router		/trips/{id}/activities [get]
func (handler *ActivityHandler) GetAllFromTrip(context *gin.Context) {
	activities, err := handler.Repository.GetAllFromTrip(context.Param("id"))
	if err != nil {
		errorHandlers.HandleGormErrors(err, context)
		return
	}

	var activitiesResponse []responses.ActivityResponse = make([]responses.ActivityResponse, len(activities))
	for i, activity := range activities {
		activitiesResponse[i] = responses.ActivityResponse{
			ID:          activity.ID,
			Name:        activity.Name,
			Description: activity.Description,
			StartDate:   activity.StartDate.Format(time.RFC3339),
			EndDate:     activity.EndDate.Format(time.RFC3339),
			Location:    activity.Location,
			Latitude:    activity.Latitude,
			Longitude:   activity.Longitude,
			Owner: responses.UserResponse{
				ID:                 activity.Owner.ID,
				Username:           activity.Owner.Username,
				ProfilePicturePath: activity.Owner.ProfilePicturePath,
				ProfilePictureUri:  activity.Owner.GetProfilePictureUri(),
			},
			Price: activity.Price,
		}
	}

	context.JSON(http.StatusOK, activitiesResponse)
	logger.ApiInfo(context, "Get all activities from trip "+context.Param("id"))
}

// @Summary		Create an activity
// @Description	Create an activity associated with the trip
// @Tags			activity
// @Accept		json
// @Produce		json
// @Security		BearerAuth
// @Param		id	path		string	true	"ID of the trip"
// @Param		activity	body	requests.ActivityCreateBody	true	"Activity information"
// @Success		201		{object}	responses.ActivityResponse
// @Failure		400		{object}	error
// @Failure		401		{object}	error
// @Router		/trips/{id}/activities [post]
func (handler *ActivityHandler) Create(context *gin.Context) {
	tripId := context.Param("id")
	tripIdUint, _ := strconv.ParseUint(tripId, 10, 64)

	var requestBody requests.ActivityCreateBody
	isBodyValid := utils.Deserialize(&requestBody, context)
	if !isBodyValid {
		return
	}

	startDate, startDateParseError := time.Parse(time.RFC3339, requestBody.StartDate)
	if startDateParseError != nil {
		logger.ApiWarning(context, "Invalid start date "+requestBody.StartDate)
		context.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "Invalid start date"})
		return
	}

	endDate, endDateParseError := time.Parse(time.RFC3339, requestBody.EndDate)
	if endDateParseError != nil {
		logger.ApiWarning(context, "Invalid end date "+requestBody.EndDate)
		context.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "Invalid end date"})
		return
	}

	currentUser, _ := utils.GetCurrentUser(context)

	var activity = models.Activity{
		Name:        requestBody.Name,
		Description: requestBody.Description,
		Trip:        models.Trip{Model: gorm.Model{ID: uint(tripIdUint)}},
		Owner:       currentUser,
		Price:       requestBody.Price,
		StartDate:   startDate,
		EndDate:     endDate,
		Location:    requestBody.Location,
	}

	err := handler.Repository.Create(&activity)
	if err != nil {
		errorHandlers.HandleGormErrors(err, context)
		return
	}

	context.JSON(http.StatusCreated, responses.ActivityResponse{
		ID:          activity.ID,
		Name:        activity.Name,
		Description: activity.Description,
		StartDate:   activity.StartDate.Format(time.RFC3339),
		EndDate:     activity.EndDate.Format(time.RFC3339),
		Price:       activity.Price,
		Location:    activity.Location,
		Latitude:    activity.Latitude,
		Longitude:   activity.Longitude,
		Owner: responses.UserResponse{
			ID:                 activity.Owner.ID,
			Username:           activity.Owner.Username,
			ProfilePicturePath: activity.Owner.ProfilePicturePath,
			ProfilePictureUri:  activity.Owner.GetProfilePictureUri(),
		},
	})
	logger.ApiInfo(context, "Activity "+activity.Name+" created on trip "+tripId)
}

// @Summary		Update an activity
// @Description	Update an activity associated with the trip
// @Tags			activity
// @Accept		json
// @Produce		json
// @Security		BearerAuth
// @Param		id	path		string	true	"ID of the trip"
// @Param		activityId	path	string	true	"ID of the activity"
// @Param		activity	body	requests.ActivityUpdateBody	true	"Activity information"
// @Success		200		{object}	responses.ActivityResponse
// @Failure		400		{object}	error
// @Failure		401		{object}	error
// @Router		/trips/{id}/activities/{activityId} [patch]
func (handler *ActivityHandler) Update(context *gin.Context) {
	activityId := context.Param("activityID")

	var requestBody requests.ActivityUpdateBody
	isBodyValid := utils.Deserialize(&requestBody, context)
	if !isBodyValid {
		return
	}

	activity, err := handler.Repository.Get(activityId)
	if err != nil {
		errorHandlers.HandleGormErrors(err, context)
		return
	}

	if requestBody.Name != "" {
		activity.Name = requestBody.Name
	}

	if requestBody.Description != "" {
		activity.Description = requestBody.Description
	}

	if requestBody.Price != 0 {
		activity.Price = requestBody.Price
	}

	if requestBody.StartDate != "" {
		startDate, startDateParseError := time.Parse(time.RFC3339, requestBody.StartDate)
		if startDateParseError != nil {
			logger.ApiWarning(context, "Invalid start date "+requestBody.StartDate)
			context.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "Invalid start date"})
			return
		}
		activity.StartDate = startDate
	}

	if requestBody.EndDate != "" {
		endDate, endDateParseError := time.Parse(time.RFC3339, requestBody.EndDate)
		if endDateParseError != nil {
			logger.ApiWarning(context, "Invalid end date "+requestBody.EndDate)
			context.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "Invalid end date"})
			return
		}
		activity.EndDate = endDate
	}

	if requestBody.Location != "" {
		activity.Location = requestBody.Location
	}

	err = handler.Repository.Update(&activity)
	if err != nil {
		fmt.Println(err)
		errorHandlers.HandleGormErrors(err, context)
		return
	}

	context.JSON(http.StatusOK, responses.ActivityResponse{
		ID:          activity.ID,
		Name:        activity.Name,
		Description: activity.Description,
		StartDate:   activity.StartDate.Format(time.RFC3339),
		EndDate:     activity.EndDate.Format(time.RFC3339),
		Price:       activity.Price,
		Location:    activity.Location,
		Latitude:    activity.Latitude,
		Longitude:   activity.Longitude,
		Owner: responses.UserResponse{
			ID:                 activity.Owner.ID,
			Username:           activity.Owner.Username,
			ProfilePicturePath: activity.Owner.ProfilePicturePath,
			ProfilePictureUri:  activity.Owner.GetProfilePictureUri(),
		},
	})
	logger.ApiInfo(context, "Activity "+activity.Name+" updated")
}

// @Summary		Delete an activity
// @Description	Delete an activity associated with the trip
// @Tags			activity
// @Accept		json
// @Produce		json
// @Security		BearerAuth
// @Param		id	path		string	true	"ID of the trip"
// @Param		activityId	path	string	true	"ID of the activity"
// @Success		204		{object}	error
// @Failure		400		{object}	error
// @Failure		401		{object}	error
// @Router		/trips/{id}/activities/{activityId} [delete]
func (handler *ActivityHandler) Delete(context *gin.Context) {
	activityId := context.Param("activityID")

	err := handler.Repository.Delete(activityId)
	if err != nil {
		errorHandlers.HandleGormErrors(err, context)
		return
	}

	context.JSON(http.StatusNoContent, nil)
	logger.ApiInfo(context, "Activity "+activityId+" deleted")
}