# go/api/requests/activity.go
package requests

type ActivityCreateBody struct {
	Name        string  `json:"name" binding:"required"`
	Description string  `json:"description"`
	Price       float64 `json:"price" binding:"required"`
	Location    string  `json:"location"`
	StartDate   string  `json:"startDate" binding:"required"`
	EndDate     string  `json:"endDate" binding:"required"`
}

type ActivityUpdateBody struct {
	Name        string  `json:"name"`
	Description string  `json:"description"`
	Price       float64 `json:"price"`
	Location    string  `json:"location"`
	StartDate   string  `json:"startDate"`
	EndDate     string  `json:"endDate"`
}